added logging of cache copy time (resolves #7)
There is now a very simple logging function for the boot time.
The first usage of it is to log how long it takes
for the prepareCacheDir method to copy all files to /tmp.

The output looks like this:

[2020-04-10T15:30:48+02:00] cache directory prepared in 82 ms.

This is also the first answer to the question how long
copying files takes.