use overlayfs (resolves #6 and probably #9)
Through this change I made the entire /var folder
writable without needing to copy anything.

It'll even work with all bundles that somehow write
into the var folder since there is no detectable
difference to a real writable folder, even when
accessed though command and overlayfs is heavily
tested since it is one of the core components of docker.